+++ lib/components/EmailInput.js
@@ -9,12 +9,14 @@ import { isEmpty } from "ramda";
 import {
   EMAIL_SEPARATION_REGEX,
   CUSTOM_STYLES,
+} from "../constants/EmailInput";
+import {
   formatEmailInputOptions,
   pruneDuplicates,
   renderValidEmails,
   renderDefaultText,
   getEmailsCount
-} from "../constants/EmailInput";
+} from "../helpers/EmailInput";
 import { hyphenize } from "../common";
 import Typography from "./Typography";
 import Label from "./Label";
